+++ angular-workspace/projects/example-client-app/src/app/login/login.component.ts
@@ -3,7 +3,7 @@ import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
 
 @Component({
-    selector: 'nimble-example-login',
+    selector: 'example-login',
     templateUrl: './login.component.html',
     styleUrls: ['./login.component.scss']
 })